requirements.txt
django
markdown
pyyaml
